+++ README.md
@@ -1,11 +1,11 @@
 <p align="center">
   <a href="https://serenitystar.ai/">
-    <h1 align="center">Serenity AIHub SDK for .NET</h1>
+    <h1 align="center">Serenity Star SDK for .NET</h1>
   </a>
 </p>
 
 <p align="center">
-  <a aria-label="try Serenity AIHub" href="https://hub.serenitystar.ai/Identity/Account/Register"><b>Try Serenity AIHub</b></a>
+  <a aria-label="try Serenity Star" href="https://hub.serenitystar.ai/Identity/Account/Register"><b>Try Serenity Star</b></a>
 &ensp;•&ensp;
   <a aria-label="Serenity documentation" href="https://docs.serenitystar.ai">Read the Documentation</a>
 &ensp;•&ensp;
@@ -16,20 +16,14 @@
 
 ## Introduction
 
-Official .NET SDK for Serenity AIHub API. This SDK provides a convenient way to interact with Serenity AIHub services from your .NET applications, allowing you to create conversations, send messages, and execute commands through a clean and intuitive interface.
+Official .NET SDK for Serenity Star API. This SDK provides a convenient way to interact with Serenity Star services from your .NET applications, allowing you to create conversations, send messages, and execute commands through a clean and intuitive interface.
 
 ## 🚀 Installation
 
-Install the Serenity AIHub SDK via NuGet:
+Install the Serenity Star SDK via NuGet:
 
 ```bash
-dotnet add package Serenity.AIHub.SDK.NET.Core
-```
-
-If you want to use dependency injection:
-
-```bash
-dotnet add package Serenity.AIHub.SDK.NET.Extensions
+dotnet add package SerenityStar.SDK.NET
 ```
 
 ## 🔧 Usage
@@ -39,11 +33,10 @@ dotnet add package Serenity.AIHub.SDK.NET.Extensions
 Create and use the client directly by providing your API key:
 
 ```csharp
-using Serenity.AIHub.SDK.NET.Core.Client;
+using SerenityStar.SDK.NET.Core.Client;
 
-var client = SerenityAIHubClient.Create("your-api-key");
+var client = SerenityClient.Create("your-api-key");
 var conversation = await client.CreateConversation("assistantagent", null);
-var response = await client.SendMessage("assistantagent", conversation.ChatId, "Hello!");
 ```
 
 ### Using Dependency Injection
@@ -52,8 +45,8 @@ Register the client in your service collection:
 
 ```csharp
 // In your startup/program.cs
-using Serenity.AIHub.SDK.NET.Extensions;
-services.AddSerenityAIHub("your-api-key");
+using SerenityStar.SDK.NET.Core.Extensions;
+services.AddSerenityStar("your-api-key");
 ```
 
 Inject and use the client in your services:
@@ -62,9 +55,9 @@ Inject and use the client in your services:
 // In your service/controller
 public class YourService
 {
-    private readonly ISerenityAIHubClient _client;
+    private readonly ISerenityClient _client;
 
-    public YourService(ISerenityAIHubClient client)
+    public YourService(ISerenityClient client)
     {
         _client = client;
     }
@@ -73,15 +66,9 @@ public class YourService
     {
         var conversation = await _client.CreateConversation("assistantagent");
 
-        List<ExecuteParameter> input = [];
-        input.Add(new(
-                "chatId",
-                conversation.ChatId
-            ));
-        input.Add(new(
-            "message",
-            "Hello, how are you?"
-        ));
+        List<ExecuteParameter> input = new List<ExecuteParameter>();
+        input.Add(new ExecuteParameter("chatId", conversation.ChatId));
+        input.Add(new ExecuteParameter("message", "Hello, how are you?"));
         var response = await _client.Execute("assistantagent", input);
     }
 }
@@ -89,7 +76,7 @@ public class YourService
 
 ## 📚 Documentation
 
-<p>Learn more about Serenity AIHub <a aria-label="serenity documentation" href="https://docs.serenitystar.ai">in our official docs!</a></p>
+<p>Learn more about Serenity Star <a aria-label="serenity documentation" href="https://docs.serenitystar.ai">in our official docs!</a></p>
 
 - [Getting Started](https://docs.serenitystar.ai/docs/getting-started/introduction)
 - [API Reference](https://docs.serenitystar.ai/docs/api/aihub/serenity-star-api-docs)
@@ -97,8 +84,7 @@ public class YourService
 
 ## 🗺 Project Structure
 
-- `src/Serenity.AIHub.SDK.NET.Core` - Core SDK package
-- `src/Serenity.AIHub.SDK.NET.Extensions` - Extension methods for dependency injection
+- `src/SerenityStar.SDK.NET.Core` - Core SDK package
 - `tests` - Test projects for the SDK
 
 ## 🤝 Contributing
