+++ SerenityStar.SDK.NET.sln
@@ -4,13 +4,11 @@ VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{13E3EF07-AD57-49FB-B936-3CD0D980368D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serenity.AIHub.SDK.NET.Core", "src\Serenity.AIHub.SDK.NET.Core\Serenity.AIHub.SDK.NET.Core.csproj", "{C77A1428-255B-4B08-A3C3-23E6FD5519D4}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serenity.AIHub.SDK.NET.Extensions", "src\Serenity.AIHub.SDK.NET.Extensions\Serenity.AIHub.SDK.NET.Extensions.csproj", "{E5AFA0B2-EC73-47E5-BE8A-C880B2E4ECEB}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SerenityStar.SDK.NET.Core", "src\SerenityStar.SDK.NET.Core\SerenityStar.SDK.NET.Core.csproj", "{C77A1428-255B-4B08-A3C3-23E6FD5519D4}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{01B46505-10BE-43F1-98A9-74DA420FE588}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serenity.AIHub.SDK.NET.IntegrationTests", "tests\Serenity.AIHub.SDK.NET.IntegrationTests\Serenity.AIHub.SDK.NET.IntegrationTests.csproj", "{E96DBD05-0F9D-4D47-9EBC-FB8F84AE1F0F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SerenityStar.SDK.NET.IntegrationTests", "tests\SerenityStar.SDK.NET.IntegrationTests\SerenityStar.SDK.NET.IntegrationTests.csproj", "{E96DBD05-0F9D-4D47-9EBC-FB8F84AE1F0F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,10 +23,6 @@ Global
 		{C77A1428-255B-4B08-A3C3-23E6FD5519D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C77A1428-255B-4B08-A3C3-23E6FD5519D4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C77A1428-255B-4B08-A3C3-23E6FD5519D4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E5AFA0B2-EC73-47E5-BE8A-C880B2E4ECEB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E5AFA0B2-EC73-47E5-BE8A-C880B2E4ECEB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E5AFA0B2-EC73-47E5-BE8A-C880B2E4ECEB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E5AFA0B2-EC73-47E5-BE8A-C880B2E4ECEB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E96DBD05-0F9D-4D47-9EBC-FB8F84AE1F0F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E96DBD05-0F9D-4D47-9EBC-FB8F84AE1F0F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E96DBD05-0F9D-4D47-9EBC-FB8F84AE1F0F}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -36,7 +30,6 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{C77A1428-255B-4B08-A3C3-23E6FD5519D4} = {13E3EF07-AD57-49FB-B936-3CD0D980368D}
-		{E5AFA0B2-EC73-47E5-BE8A-C880B2E4ECEB} = {13E3EF07-AD57-49FB-B936-3CD0D980368D}
 		{E96DBD05-0F9D-4D47-9EBC-FB8F84AE1F0F} = {01B46505-10BE-43F1-98A9-74DA420FE588}
 	EndGlobalSection
 EndGlobal
