--- src/Serenity.AIHub.SDK.NET.Core/Constants/ClientConstants.cs
@@ -1,12 +1,0 @@
-namespace Serenity.AIHub.SDK.NET.Core.Constants;
-
-/// <summary>
-/// Contains constant values used throughout the SDK.
-/// </summary>
-public static class ClientConstants
-{
-    /// <summary>
-    /// The base URL for the Serenity AI Hub API.
-    /// </summary>
-    public const string BaseUrl = "https://api.serenitystar.ai/api";
-}
