--- src/Serenity.AIHub.SDK.NET.Core/Serenity.AIHub.SDK.NET.Core.csproj
@@ -1,29 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworks>net6.0;net7.0;net8.0</TargetFrameworks>
-    <PackageId>Serenity.AIHub.SDK.NET.Core</PackageId>
-    <Authors>Matías Rohleder</Authors>
-    <Company>Subgen AI</Company>
-    <Description>Official .NET SDK for Serenity AI Hub API</Description>
-    <PackageProjectUrl>https://github.com/matiasrohleder/Serenity.AIHub.SDK.NET</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/matiasrohleder/Serenity.AIHub.SDK.NET</RepositoryUrl>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageReadmeFile>README.md</PackageReadmeFile>
-    <PackageTags>ai;chat;sdk;serenity;aihub</PackageTags>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <PackageIcon>icon.png</PackageIcon>
-    <Version>1.1.6</Version>
-    <LangVersion>12.0</LangVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Include="..\..\README.md" Pack="true" PackagePath="/" />
-    <None Include="..\..\icon.png" Pack="true" PackagePath="/" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Http" Version="6.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Options" Version="6.0.0" />
-  </ItemGroup>
-</Project>
