--- src/Serenity.AIHub.SDK.NET.Extensions/Serenity.AIHub.SDK.NET.Extensions.csproj
@@ -1,31 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworks>net6.0;net7.0;net8.0</TargetFrameworks>
-    <PackageId>Serenity.AIHub.SDK.NET.Extensions</PackageId>
-    <Authors>Matías Rohleder</Authors>
-    <Company>Subgen AI</Company>
-    <Description>Extensions for Serenity AI Hub SDK</Description>
-    <PackageProjectUrl>https://github.com/matiasrohleder/Serenity.AIHub.SDK.NET</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/matiasrohleder/Serenity.AIHub.SDK.NET</RepositoryUrl>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageReadmeFile>README.md</PackageReadmeFile>
-    <PackageTags>ai;chat;sdk;serenity;aihub;extensions</PackageTags>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <PackageIcon>icon.png</PackageIcon>
-    <Version>1.1.6</Version>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Serenity.AIHub.SDK.NET.Core\Serenity.AIHub.SDK.NET.Core.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="..\..\README.md" Pack="true" PackagePath="/" />
-    <None Include="..\..\icon.png" Pack="true" PackagePath="/" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="6.0.0" />
-  </ItemGroup>
-</Project>
