+++ src/SerenityStar.SDK.NET.Core/Constants/ClientConstants.cs
@@ -1,0 +1,13 @@
+namespace SerenityStar.SDK.NET.Core.Constants
+{
+    /// <summary>
+    /// Contains constant values used throughout the SDK.
+    /// </summary>
+    internal static class ClientConstants
+    {
+        /// <summary>
+        /// The base URL for the Serenity AI Hub API.
+        /// </summary>
+        internal const string BaseUrl = "https://api.serenitystar.ai/api";
+    }
+}
