--- tests/Serenity.AIHub.SDK.NET.IntegrationTests/appsettings.json
@@ -1,5 +1,0 @@
-{
-    "SerenityAIHub": {
-        "ApiKey": "your-api-key-here"
-    }
-}
