+++ tests/SerenityStar.SDK.NET.IntegrationTests/appsettings.json
@@ -1,0 +1,7 @@
+{
+    "SerenityStar": {
+        "ApiKey": "your-api-key-here",
+        "AssistantAgent": "assistantagent",
+        "ActivityAgent": "activityagent"
+    }
+}
